@@ -40,8 +40,8 @@ test {
 }
 
 compileTestJava {
-    sourceCompatibility = 1.8
-    targetCompatibility = 1.8
+    sourceCompatibility = 17
+    targetCompatibility = 17
     options.compilerArgs += '-parameters'
 }
 
